Add react spring loading
