+++ package.json
@@ -20,6 +20,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "2.0.0",
+    "react-spring": "^8.0.27",
     "styled-components": "5.0.1"
   }
 }
